@@ -21,83 +21,83 @@
   "bugs": {
     "url": "https://github.com/openlattice/lattice-edm/issues"
   },
-  "license": "GPLv3",
+  "license": "GPL-3.0",
   "dependencies": {
-    "@babel/polyfill": "^7.0.0",
-    "@babel/runtime": "^7.0.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.0",
-    "@fortawesome/pro-light-svg-icons": "^5.3.0",
-    "@fortawesome/pro-regular-svg-icons": "^5.3.0",
-    "@fortawesome/pro-solid-svg-icons": "^5.3.0",
-    "@fortawesome/react-fontawesome": "^0.1.0",
-    "history": "^4.7.2",
+    "@babel/polyfill": "~7.2.0",
+    "@babel/runtime": "~7.3.0",
+    "@fortawesome/fontawesome-svg-core": "~1.2.0",
+    "@fortawesome/pro-light-svg-icons": "~5.7.0",
+    "@fortawesome/pro-regular-svg-icons": "~5.7.0",
+    "@fortawesome/pro-solid-svg-icons": "~5.7.0",
+    "@fortawesome/react-fontawesome": "~0.1.0",
+    "@redux-saga/core": "~1.0.0",
+    "axios": "~0.18.0",
+    "connected-react-router": "~6.3.0",
+    "date-fns": "~1.30.0",
+    "history": "~4.9.0",
     "immutable": "4.0.0-rc.9",
-    "lattice": "^0.44.0",
-    "lattice-auth": "^0.9.1",
-    "lattice-sagas": "^0.10.0",
-    "lodash": "^4.17.10",
-    "loglevel": "^1.6.0",
-    "moment": "^2.22.0",
-    "polished": "^2.0.1",
-    "qs": "^6.5.1",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^5.0.0-alpha.9",
-    "react-sortable-hoc": "^0.8.0",
-    "react-virtualized": "^9.20.0",
-    "redux": "^4.0.0",
-    "redux-immutable": "^4.0.0",
-    "redux-reqseq": "0.1.1",
-    "redux-saga": "^0.16.0",
-    "styled-components": "^3.4.0",
-    "validator": "^10.7.0"
+    "lattice": "0.55.0",
+    "lattice-auth": "0.12.2",
+    "lattice-sagas": "0.16.0",
+    "lodash": "~4.17.10",
+    "loglevel": "~1.6.0",
+    "polished": "~3.2.0",
+    "qs": "~6.7.0",
+    "react": "~16.8.0",
+    "react-dom": "~16.8.0",
+    "react-redux": "~6.0.0",
+    "react-router": "~5.0.0",
+    "react-router-dom": "~5.0.0",
+    "react-sortable-hoc": "~1.8.0",
+    "react-virtualized": "~9.21.0",
+    "redux": "~4.0.0",
+    "redux-immutable": "4.0.0",
+    "redux-reqseq": "0.2.1",
+    "styled-components": "~4.2.0",
+    "validator": "~10.11.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.0.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.0",
-    "babel-jest": "^23.4.2",
-    "babel-loader": "^8.0.0",
-    "babel-plugin-styled-components": "^1.6.0",
-    "enzyme": "^3.6.0",
-    "enzyme-adapter-react-16": "^1.4.0",
-    "enzyme-to-json": "^3.3.4",
-    "eslint": "^5.5.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-import-resolver-webpack": "^0.10.0",
-    "eslint-loader": "^2.1.0",
-    "eslint-plugin-flowtype": "^3.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^21.22.0",
-    "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-react": "^7.11.0",
-    "file-loader": "^2.0.0",
-    "flow-bin": "^0.84.0",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^23.5.0",
-    "jest-styled-components": "^6.1.0",
-    "npm-run-all": "^4.1.0",
-    "rimraf": "^2.6.0",
-    "stylelint": "^9.5.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.3.0",
-    "uuid": "^3.3.0",
-    "webpack": "^4.17.0",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.0",
-    "webpack-hot-middleware": "^2.23.0"
+    "@babel/cli": "~7.2.0",
+    "@babel/core": "~7.2.0",
+    "@babel/plugin-proposal-class-properties": "~7.2.0",
+    "@babel/plugin-proposal-object-rest-spread": "~7.2.0",
+    "@babel/plugin-transform-runtime": "~7.2.0",
+    "@babel/preset-env": "~7.2.0",
+    "@babel/preset-flow": "~7.0.0",
+    "@babel/preset-react": "~7.0.0",
+    "@babel/register": "~7.0.0",
+    "babel-eslint": "~10.0.0",
+    "babel-jest": "~24.1.0",
+    "babel-loader": "~8.0.0",
+    "babel-plugin-styled-components": "~1.10.0",
+    "enzyme": "~3.9.0",
+    "enzyme-adapter-react-16": "~1.10.0",
+    "enzyme-to-json": "~3.3.5",
+    "eslint": "~5.16.0",
+    "eslint-config-airbnb": "~17.1.0",
+    "eslint-import-resolver-webpack": "~0.11.0",
+    "eslint-loader": "~2.1.0",
+    "eslint-plugin-flowtype": "~3.6.0",
+    "eslint-plugin-import": "~2.16.0",
+    "eslint-plugin-jest": "~22.4.0",
+    "eslint-plugin-jsx-a11y": "~6.2.0",
+    "eslint-plugin-react": "~7.12.0",
+    "file-loader": "~3.0.0",
+    "flow-bin": "~0.96.0",
+    "html-webpack-plugin": "~3.2.0",
+    "jest": "~24.1.0",
+    "jest-styled-components": "~6.3.0",
+    "npm-run-all": "~4.1.0",
+    "rimraf": "~2.6.0",
+    "stylelint": "~9.10.0",
+    "stylelint-config-standard": "~18.2.0",
+    "stylelint-config-styled-components": "~0.1.1",
+    "stylelint-processor-styled-components": "~1.6.0",
+    "uuid": "~3.3.0",
+    "webpack": "~4.29.0",
+    "webpack-cli": "~3.3.0",
+    "webpack-dev-server": "~3.3.0",
+    "webpack-hot-middleware": "~2.24.0"
   },
   "scripts": {
     "app": "webpack-dev-server --config config/webpack/webpack.config.app.js --env.development",
