@@ -1,4 +1,3 @@
-const moment = require('moment');
 const PACKAGE = require('../../package.json');
 
 const BANNER = `
@@ -6,7 +5,7 @@ ${PACKAGE.name} - v${PACKAGE.version}
 ${PACKAGE.description}
 ${PACKAGE.homepage}
 
-Copyright (c) 2017-${moment().year()}, OpenLattice, Inc. All rights reserved.
+Copyright (c) 2017-${(new Date()).getFullYear()}, OpenLattice, Inc. All rights reserved.
 `;
 
 module.exports = {
