@@ -1,10 +1,10 @@
-// flow-typed signature: 7a63f58e89f96a52bfb4d1adc829f41d
-// flow-typed version: <<STUB>>/lattice_v0.37.0/flow_v0.80.0
+// flow-typed signature: 09320937a85401306a3ec3c090feb5c6
+// flow-typed version: <<STUB>>/connected-react-router_v6.3.1/flow_v0.93.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'lattice'
+ *   'connected-react-router'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,10 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'lattice' {
+declare module 'connected-react-router' {
+  declare module.exports: any;
+}
+
+declare module 'connected-react-router/immutable' {
   declare module.exports: any;
 }
