@@ -7,38 +7,36 @@ import React from 'react';
 import styled from 'styled-components';
 import { List, Map, Set } from 'immutable';
 import { Models, Types } from 'lattice';
-import { EntityDataModelApiActionFactory } from 'lattice-sagas';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
+import type { AnalyzerType, FQN } from 'lattice';
 
 import AbstractTypes from '../../utils/AbstractTypes';
 import AbstractTypeDataTable from '../../components/datatable/AbstractTypeDataTable';
 import AbstractTypeSearchableSelect from '../../components/controls/AbstractTypeSearchableSelect';
 import InlineEditableControl from '../../components/controls/InlineEditableControl';
+import Logger from '../../utils/Logger';
 import StyledButton from '../../components/buttons/StyledButton';
 import StyledCard from '../../components/cards/StyledCard';
-
+import * as AssociationTypesActions from './associationtypes/AssociationTypesActions';
+import * as EntityTypesActions from './entitytypes/EntityTypesActions';
+import * as PropertyTypesActions from './propertytypes/PropertyTypesActions';
+import * as SchemasActions from './schemas/SchemasActions';
 import { EDM_PRIMITIVE_TYPES } from '../../utils/EdmPrimitiveTypes';
-
 import type { AbstractType } from '../../utils/AbstractTypes';
 
-const {
-  createAssociationType,
-  createEntityType,
-  createPropertyType,
-  createSchema
-} = EntityDataModelApiActionFactory;
+const LOG :Logger = new Logger('AbstractTypeCreateContainer');
 
 const {
-  FullyQualifiedName,
   AssociationType,
   AssociationTypeBuilder,
   EntityType,
   EntityTypeBuilder,
+  FullyQualifiedName,
   PropertyType,
   PropertyTypeBuilder,
   Schema,
-  SchemaBuilder
+  SchemaBuilder,
 } = Models;
 
 const {
@@ -64,11 +62,17 @@ const DATA_TYPE_OPTIONS = EDM_PRIMITIVE_TYPES.map((primitive :string) => (
   </option>
 ));
 
+const ANALYZER_TYPE_OPTIONS = Object.keys(AnalyzerTypes).map((anaylzer :AnalyzerType) => (
+  <option key={anaylzer} value={anaylzer}>
+    { anaylzer }
+  </option>
+));
+
 /*
  * styled components
  */
 
-const AbstractTypeCreateCard = StyledCard.extend`
+const AbstractTypeCreateCard = styled(StyledCard)`
   flex: 50%;
   max-width: 1000px;
   min-width: 500px;
@@ -85,7 +89,7 @@ const ActionButtons = styled.div`
   }
 `;
 
-const DataTypeSelect = styled.select`
+const TypeSelect = styled.select`
   align-self: left;
   background: none;
   border: 1px solid #c5d5e5;
@@ -129,18 +133,20 @@ const PropertyTypesSection = styled.section`
 
 type Props = {
   actions :{
-    createAssociationType :RequestSequence;
-    createEntityType :RequestSequence;
-    createPropertyType :RequestSequence;
-    createSchema :RequestSequence;
+    localCreateAssociationType :RequestSequence;
+    localCreateEntityType :RequestSequence;
+    localCreatePropertyType :RequestSequence;
+    localCreateSchema :RequestSequence;
   };
   entityTypes :List<Map<*, *>>;
   propertyTypes :List<Map<*, *>>;
   workingAbstractTypeType :AbstractType;
   onCancel :() => void;
+  onSubmit :() => void;
 };
 
 type State = {
+  analyzerValue :AnalyzerType;
   bidiValue :boolean;
   datatypeValue :string;
   descriptionValue :string;
@@ -160,16 +166,12 @@ type State = {
 
 class AbstractTypeCreateContainer extends React.Component<Props, State> {
 
-  static defaultProps = {
-    workingAbstractTypeType: AbstractTypes.PropertyType,
-    onCancel: () => {}
-  }
-
   constructor(props :Props) {
 
     super(props);
 
     this.state = {
+      analyzerValue: AnalyzerTypes.STANDARD,
       bidiValue: false,
       datatypeValue: 'String',
       descriptionValue: '',
@@ -184,7 +186,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
       selectedPrimaryKeyPropertyTypes: Set(),
       selectedPropertyTypes: Set(),
       selectedSourceEntityTypes: Set(),
-      titleValue: ''
+      titleValue: '',
     };
   }
 
@@ -231,8 +233,9 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
 
   submitCreateAbstractTypeRequest = () => {
 
-    const { actions, workingAbstractTypeType } = this.props;
+    const { actions, onSubmit, workingAbstractTypeType } = this.props;
     const {
+      analyzerValue,
       bidiValue,
       datatypeValue,
       descriptionValue,
@@ -253,13 +256,13 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
         .setFullyQualifiedName(new FullyQualifiedName(namespaceValue, nameValue))
         .build();
 
-      actions.createSchema(newSchema);
+      actions.localCreateSchema(newSchema);
     }
     else if (workingAbstractTypeType === AbstractTypes.PropertyType) {
 
-      const analyzer :string = (datatypeValue === 'String' && phoneticSearchesValue)
+      const analyzer :AnalyzerType = (datatypeValue === 'String' && phoneticSearchesValue)
         ? AnalyzerTypes.METAPHONE
-        : AnalyzerTypes.STANDARD;
+        : analyzerValue;
 
       const newPropertyType :PropertyType = new PropertyTypeBuilder()
         .setType(new FullyQualifiedName(namespaceValue, nameValue))
@@ -270,7 +273,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
         .setAnalyzer(analyzer)
         .build();
 
-      actions.createPropertyType(newPropertyType);
+      actions.localCreatePropertyType(newPropertyType);
     }
     else {
 
@@ -294,7 +297,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
           .setCategory(SecurableTypes.EntityType)
           .build();
 
-        actions.createEntityType(newEntityType);
+        actions.localCreateEntityType(newEntityType);
       }
       else if (workingAbstractTypeType === AbstractTypes.AssociationType) {
 
@@ -315,51 +318,54 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
           .setBidirectional(bidiValue)
           .build();
 
-        actions.createAssociationType(newAssociationType);
+        actions.localCreateAssociationType(newAssociationType);
       }
       else {
-        // TODO: need a Logger class
-        // console.error('invalid AbstractType: ', this.props.workingAbstractTypeType);
+        LOG.error('invalid AbstractType:', workingAbstractTypeType);
       }
+    }
+
+    if (onSubmit) {
+      onSubmit();
     }
   }
 
-  addToSelectedDestinationEntityTypes = (selectedEntityTypeId :string) => {
+  addToSelectedDestinationEntityTypes = (selectedEntityTypeFQN :FQN) => {
 
     const { entityTypes } = this.props;
     const { selectedDestinationEntityTypes } = this.state;
 
     const selectedEntityType :Map<*, *> = entityTypes.find((entityType :Map<*, *>) => (
-      entityType.get('id', '') === selectedEntityTypeId
+      FullyQualifiedName.toString(entityType.get('type', Map())) === selectedEntityTypeFQN.toString()
     ));
 
     this.setState({
-      selectedDestinationEntityTypes: selectedDestinationEntityTypes.add(selectedEntityType)
+      selectedDestinationEntityTypes: selectedDestinationEntityTypes.add(selectedEntityType),
     });
   }
 
-  addToSelectedPrimaryKeyPropertyTypes = (selectedPropertyTypeId :string) => {
+  addToSelectedPrimaryKeyPropertyTypes = (selectedPropertyTypeFQN :FQN) => {
 
     const { propertyTypes } = this.props;
     const { selectedPrimaryKeyPropertyTypes, selectedPropertyTypes } = this.state;
 
     const selectedPropertyType :Map<*, *> = propertyTypes.find((propertyType :Map<*, *>) => (
-      propertyType.get('id', '') === selectedPropertyTypeId
+      FullyQualifiedName.toString(propertyType.get('type', Map())) === selectedPropertyTypeFQN.toString()
     ));
 
     this.setState({
       selectedPrimaryKeyPropertyTypes: selectedPrimaryKeyPropertyTypes.add(selectedPropertyType),
-      selectedPropertyTypes: selectedPropertyTypes.subtract(selectedPrimaryKeyPropertyTypes)
+      selectedPropertyTypes: selectedPropertyTypes.subtract(selectedPrimaryKeyPropertyTypes),
     });
   }
 
-  addToSelectedPropertyTypes = (selectedPropertyTypeId :string) => {
+  addToSelectedPropertyTypes = (selectedPropertyTypeFQN :FQN) => {
 
     const { propertyTypes } = this.props;
     const { selectedPrimaryKeyPropertyTypes, selectedPropertyTypes } = this.state;
 
     const selectedPropertyType :Map<*, *> = propertyTypes.find((propertyType :Map<*, *>) => (
-      propertyType.get('id', '') === selectedPropertyTypeId
+      FullyQualifiedName.toString(propertyType.get('type', Map())) === selectedPropertyTypeFQN.toString()
     ));
 
     if (selectedPrimaryKeyPropertyTypes.contains(selectedPropertyType)) {
@@ -367,90 +373,113 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
     }
 
     this.setState({
-      selectedPropertyTypes: selectedPropertyTypes.add(selectedPropertyType)
+      selectedPropertyTypes: selectedPropertyTypes.add(selectedPropertyType),
     });
   }
 
-  addToSelectedSourceEntityTypes = (selectedEntityTypeId :string) => {
+  addToSelectedSourceEntityTypes = (selectedEntityTypeFQN :FQN) => {
 
     const { entityTypes } = this.props;
     const { selectedSourceEntityTypes } = this.state;
 
     const selectedEntityType :Map<*, *> = entityTypes.find((entityType :Map<*, *>) => (
-      entityType.get('id', '') === selectedEntityTypeId
+      FullyQualifiedName.toString(entityType.get('type', Map())) === selectedEntityTypeFQN.toString()
     ));
 
     this.setState({
-      selectedSourceEntityTypes: selectedSourceEntityTypes.add(selectedEntityType)
+      selectedSourceEntityTypes: selectedSourceEntityTypes.add(selectedEntityType),
     });
   }
 
-  removeFromSelectedDestinationEntityTypes = (selectedEntityTypeId :string) => {
+  removeFromSelectedDestinationEntityTypes = (selectedEntityTypeFQN :FQN) => {
 
     const { selectedDestinationEntityTypes } = this.state;
 
     const updatedSelectedDestinationEntityTypes :Set<Map<*, *>> = selectedDestinationEntityTypes
-      .filterNot((entityType :Map<*, *>) => entityType.get('id', '') === selectedEntityTypeId);
+      .filterNot((entityType :Map<*, *>) => (
+        FullyQualifiedName.toString(entityType.get('type', Map())) === selectedEntityTypeFQN.toString()
+      ));
 
     this.setState({
-      selectedDestinationEntityTypes: updatedSelectedDestinationEntityTypes
+      selectedDestinationEntityTypes: updatedSelectedDestinationEntityTypes,
     });
   }
 
-  removeFromSelectedPrimaryKeyPropertyTypes = (selectedPropertyTypeId :string) => {
+  removeFromSelectedPrimaryKeyPropertyTypes = (selectedPropertyTypeFQN :FQN) => {
 
     const { selectedPrimaryKeyPropertyTypes } = this.state;
 
     const updatedSelectedPrimaryKeyPropertyTypes :Set<Map<*, *>> = selectedPrimaryKeyPropertyTypes
-      .filterNot((propertyType :Map<*, *>) => propertyType.get('id', '') === selectedPropertyTypeId);
+      .filterNot((propertyType :Map<*, *>) => (
+        FullyQualifiedName.toString(propertyType.get('type', Map())) === selectedPropertyTypeFQN.toString()
+      ));
 
     this.setState({
-      selectedPrimaryKeyPropertyTypes: updatedSelectedPrimaryKeyPropertyTypes
+      selectedPrimaryKeyPropertyTypes: updatedSelectedPrimaryKeyPropertyTypes,
     });
   }
 
-  removeFromSelectedPropertyTypes = (selectedPropertyTypeId :string) => {
+  removeFromSelectedPropertyTypes = (selectedPropertyTypeFQN :FQN) => {
 
     const { selectedPropertyTypes } = this.state;
 
     const updatedSelectedPropertyTypes :Set<Map<*, *>> = selectedPropertyTypes
-      .filterNot((propertyType :Map<*, *>) => propertyType.get('id', '') === selectedPropertyTypeId);
+      .filterNot((propertyType :Map<*, *>) => (
+        FullyQualifiedName.toString(propertyType.get('type', Map())) === selectedPropertyTypeFQN.toString()
+      ));
 
     this.setState({
-      selectedPropertyTypes: updatedSelectedPropertyTypes
+      selectedPropertyTypes: updatedSelectedPropertyTypes,
     });
   }
 
-  removeFromSelectedSourceEntityTypes = (selectedEntityTypeId :string) => {
+  removeFromSelectedSourceEntityTypes = (selectedEntityTypeFQN :FQN) => {
 
     const { selectedSourceEntityTypes } = this.state;
 
     const updatedSelectedSourceEntityTypes :Set<Map<*, *>> = selectedSourceEntityTypes
-      .filterNot((entityType :Map<*, *>) => entityType.get('id', '') === selectedEntityTypeId);
+      .filterNot((entityType :Map<*, *>) => (
+        FullyQualifiedName.toString(entityType.get('type', Map())) === selectedEntityTypeFQN.toString()
+      ));
 
     this.setState({
-      selectedSourceEntityTypes: updatedSelectedSourceEntityTypes
+      selectedSourceEntityTypes: updatedSelectedSourceEntityTypes,
+    });
+  }
+
+  handleOnChangeAnalyzerType = (event :SyntheticInputEvent<*>) => {
+
+    this.setState({
+      analyzerValue: event.target.value || AnalyzerTypes.STANDARD,
     });
   }
 
   handleOnChangeBidi = (event :SyntheticInputEvent<*>) => {
 
     this.setState({
-      bidiValue: event.target.id === BIDI_YES_RADIO_ID
+      bidiValue: event.target.id === BIDI_YES_RADIO_ID,
     });
   }
 
   handleOnChangeDataType = (event :SyntheticInputEvent<*>) => {
 
+    let { phoneticSearchesValue } = this.state;
+    const datatypeValue = event.target.value || 'String';
+
+    if (datatypeValue !== 'String') {
+      phoneticSearchesValue = false;
+    }
+
     this.setState({
-      datatypeValue: event.target.value || ''
+      datatypeValue,
+      phoneticSearchesValue,
     });
   }
 
   handleOnChangeDescription = (description :string) => {
 
     this.setState({
-      descriptionValue: description || ''
+      descriptionValue: description || '',
     });
   }
 
@@ -458,7 +487,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
 
     this.setState({
       nameValue: name || '',
-      isInEditModeName: false
+      isInEditModeName: false,
     });
   }
 
@@ -466,21 +495,28 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
 
     this.setState({
       namespaceValue: namespace || '',
-      isInEditModeNamespace: false
+      isInEditModeNamespace: false,
     });
   }
 
   handleOnChangePhoneticSearches = (event :SyntheticInputEvent<*>) => {
 
+    const { analyzerValue, datatypeValue } = this.state;
+    const phoneticSearchesValue = event.target.checked || false;
+    const analyzerType = (datatypeValue === 'String' && phoneticSearchesValue)
+      ? AnalyzerTypes.METAPHONE
+      : analyzerValue;
+
     this.setState({
-      phoneticSearchesValue: event.target.checked || false
+      phoneticSearchesValue,
+      analyzerValue: analyzerType,
     });
   }
 
   handleOnChangePii = (event :SyntheticInputEvent<*>) => {
 
     this.setState({
-      piiValue: event.target.id === PII_YES_RADIO_ID
+      piiValue: event.target.id === PII_YES_RADIO_ID,
     });
   }
 
@@ -488,28 +524,28 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
 
     this.setState({
       titleValue: title || '',
-      isInEditModeTitle: false
+      isInEditModeTitle: false,
     });
   }
 
   handleOnEditToggleName = (editable :boolean) => {
 
     this.setState({
-      isInEditModeName: editable
+      isInEditModeName: editable,
     });
   }
 
   handleOnEditToggleNamespace = (editable :boolean) => {
 
     this.setState({
-      isInEditModeNamespace: editable
+      isInEditModeNamespace: editable,
     });
   }
 
   handleOnEditToggleTitle = (editable :boolean) => {
 
     this.setState({
-      isInEditModeTitle: editable
+      isInEditModeTitle: editable,
     });
   }
 
@@ -545,7 +581,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
             'text',
             'Title',
             this.handleOnChangeTitle,
-            this.handleOnEditToggleTitle
+            this.handleOnEditToggleTitle,
           )
         }
       </section>
@@ -568,7 +604,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
             'textarea',
             'Description',
             this.handleOnChangeDescription,
-            () => {}
+            () => {},
           )
         }
       </section>
@@ -589,9 +625,9 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
         <h2>
           Data Type
         </h2>
-        <DataTypeSelect onChange={this.handleOnChangeDataType} defaultValue="String">
+        <TypeSelect onChange={this.handleOnChangeDataType} defaultValue="String">
           { DATA_TYPE_OPTIONS }
-        </DataTypeSelect>
+        </TypeSelect>
         {
           datatypeValue !== 'String'
             ? null
@@ -606,6 +642,26 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
               </PhoneticCheckboxWrapper>
             )
         }
+      </section>
+    );
+  }
+
+  renderPropertyTypeAnalyzerTypeSelectSection = () => {
+
+    const { workingAbstractTypeType } = this.props;
+    const { analyzerValue } = this.state;
+    if (workingAbstractTypeType !== AbstractTypes.PropertyType) {
+      return null;
+    }
+
+    return (
+      <section>
+        <h2>
+          Analyzer Type
+        </h2>
+        <TypeSelect onChange={this.handleOnChangeAnalyzerType} value={analyzerValue}>
+          { ANALYZER_TYPE_OPTIONS }
+        </TypeSelect>
       </section>
     );
   }
@@ -744,7 +800,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
             availablePropertyTypes.toList(),
             AbstractTypes.PropertyType,
             'Available PropertyTypes...',
-            this.addToSelectedPrimaryKeyPropertyTypes
+            this.addToSelectedPrimaryKeyPropertyTypes,
           )
         }
       </PrimaryKeyPropertyTypesSection>
@@ -833,7 +889,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
             entityTypes,
             AbstractTypes.EntityType,
             'Available EntityTypes...',
-            this.addToSelectedSourceEntityTypes
+            this.addToSelectedSourceEntityTypes,
           )
         }
       </section>
@@ -875,7 +931,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
             entityTypes,
             AbstractTypes.EntityType,
             'Available EntityTypes...',
-            this.addToSelectedDestinationEntityTypes
+            this.addToSelectedDestinationEntityTypes,
           )
         }
       </section>
@@ -918,7 +974,7 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
               'text',
               'Namespace',
               this.handleOnChangeNamespace,
-              this.handleOnEditToggleNamespace
+              this.handleOnEditToggleNamespace,
             )
           }
         </section>
@@ -928,13 +984,14 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
               'text',
               'Name',
               this.handleOnChangeName,
-              this.handleOnEditToggleName
+              this.handleOnEditToggleName,
             )
           }
         </section>
         { this.renderTitleSection() }
         { this.renderDescriptionSection() }
         { this.renderPropertyTypeDataTypeSelectSection() }
+        { this.renderPropertyTypeAnalyzerTypeSelectSection() }
         { this.renderPropertyTypePiiSection() }
         { this.renderAssociationTypeBidirectionalSection() }
         { this.renderEntityTypePrimaryKeyPropertyTypesSelectSection() }
@@ -964,29 +1021,19 @@ class AbstractTypeCreateContainer extends React.Component<Props, State> {
   }
 }
 
-function mapStateToProps(state :Map<*, *>) :Object {
+const mapStateToProps = (state :Map<*, *>) :{} => ({
+  associationTypes: state.getIn(['edm', 'associationTypes', 'associationTypes']),
+  entityTypes: state.getIn(['edm', 'entityTypes', 'entityTypes']),
+  propertyTypes: state.getIn(['edm', 'propertyTypes', 'propertyTypes']),
+});
 
-  return {
-    associationTypes: state.getIn(['edm', 'associationTypes', 'associationTypes']),
-    entityTypes: state.getIn(['edm', 'entityTypes', 'entityTypes']),
-    entityTypesById: state.getIn(['edm', 'entityTypes', 'entityTypesById']),
-    propertyTypes: state.getIn(['edm', 'propertyTypes', 'propertyTypes']),
-    propertyTypesById: state.getIn(['edm', 'propertyTypes', 'propertyTypesById'])
-  };
-}
-
-function mapDispatchToProps(dispatch :Function) :Object {
-
-  const actions = {
-    createAssociationType,
-    createEntityType,
-    createPropertyType,
-    createSchema
-  };
-
-  return {
-    actions: bindActionCreators(actions, dispatch)
-  };
-}
+const mapDispatchToProps = (dispatch :Function) :{} => ({
+  actions: bindActionCreators({
+    localCreateAssociationType: AssociationTypesActions.localCreateAssociationType,
+    localCreateEntityType: EntityTypesActions.localCreateEntityType,
+    localCreatePropertyType: PropertyTypesActions.localCreatePropertyType,
+    localCreateSchema: SchemasActions.localCreateSchema,
+  }, dispatch)
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(AbstractTypeCreateContainer);
