@@ -6,10 +6,11 @@ import React from 'react';
 
 import styled from 'styled-components';
 import { List, Map, OrderedSet } from 'immutable';
+import { Models } from 'lattice';
 import { AuthUtils } from 'lattice-auth';
-import { EntityDataModelApiActionFactory } from 'lattice-sagas';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
+import type { FQN } from 'lattice';
 
 import AbstractTypes from '../../../utils/AbstractTypes';
 import AbstractTypeDataTable from '../../../components/datatable/AbstractTypeDataTable';
@@ -17,20 +18,21 @@ import AbstractTypeFieldDescription from '../AbstractTypeFieldDescription';
 import AbstractTypeFieldTitle from '../AbstractTypeFieldTitle';
 import AbstractTypeFieldType from '../AbstractTypeFieldType';
 import AbstractTypeSearchableSelect from '../../../components/controls/AbstractTypeSearchableSelect';
+import Logger from '../../../utils/Logger';
 import StyledButton from '../../../components/buttons/StyledButton';
+import * as EntityTypesActions from './EntityTypesActions';
+import { isValidUUID } from '../../../utils/ValidationUtils';
+import type { IndexMap } from '../Types';
 
-const {
-  addPropertyTypeToEntityType,
-  deleteEntityType,
-  removePropertyTypeFromEntityType,
-  reorderEntityTypePropertyTypes
-} = EntityDataModelApiActionFactory;
+const LOG :Logger = new Logger('EntityTypeDetailsContainer');
+
+const { FullyQualifiedName } = Models;
 
 /*
  * styled components
  */
 
-const DeleteButton = StyledButton.extend`
+const DeleteButton = styled(StyledButton)`
   align-self: center;
 `;
 
@@ -44,65 +46,88 @@ const AbstractTypeSearchableSelectWrapper = styled.div`
 
 type Props = {
   actions :{
-    addPropertyTypeToEntityType :RequestSequence;
-    deleteEntityType :RequestSequence;
-    removePropertyTypeFromEntityType :RequestSequence;
-    reorderEntityTypePropertyTypes :RequestSequence;
+    localAddPropertyTypeToEntityType :RequestSequence;
+    localDeleteEntityType :RequestSequence;
+    localRemovePropertyTypeFromEntityType :RequestSequence;
   };
   entityType :Map<*, *>;
   propertyTypes :List<Map<*, *>>;
-  propertyTypesById :Map<string, number>;
+  propertyTypesIndexMap :IndexMap;
 }
 
 class EntityTypeDetailsContainer extends React.Component<Props> {
 
-  handleOnPropertyTypeAdd = (selectedPropertyTypeId :string) => {
+  handleOnPropertyTypeAdd = (selectedPropertyTypeFQN :FQN) => {
 
-    const { actions, entityType } = this.props;
+    const {
+      actions,
+      entityType,
+      propertyTypes,
+      propertyTypesIndexMap,
+    } = this.props;
 
     if (AuthUtils.isAuthenticated() && AuthUtils.isAdmin()) {
-      actions.addPropertyTypeToEntityType({
+      const propertyTypesIndex :number = propertyTypesIndexMap.get(selectedPropertyTypeFQN, -1);
+      const propertyTypeId :?UUID = propertyTypes.getIn([propertyTypesIndex, 'id']);
+      if (!isValidUUID(propertyTypeId)) {
+        const errorMsg = 'PropertyType id must be a valid UUID, otherwise it cannot be added to an EntityType';
+        LOG.error(errorMsg, propertyTypeId);
+        return;
+      }
+      actions.localAddPropertyTypeToEntityType({
+        propertyTypeId,
+        entityTypeFQN: new FullyQualifiedName(entityType.get('type')),
         entityTypeId: entityType.get('id'),
-        propertyTypeId: selectedPropertyTypeId
       });
     }
   }
 
-  handleOnPropertyTypeRemove = (removedAbstractTypeId :string) => {
+  handleOnPropertyTypeRemove = (removedPropertyTypeFQN :FQN) => {
 
-    const { actions, entityType } = this.props;
+    const {
+      actions,
+      entityType,
+      propertyTypes,
+      propertyTypesIndexMap,
+    } = this.props;
 
     if (AuthUtils.isAuthenticated() && AuthUtils.isAdmin()) {
-      actions.removePropertyTypeFromEntityType({
+      const propertyTypesIndex :number = propertyTypesIndexMap.get(removedPropertyTypeFQN, -1);
+      const propertyTypeId :?UUID = propertyTypes.getIn([propertyTypesIndex, 'id']);
+      actions.localRemovePropertyTypeFromEntityType({
+        propertyTypeId,
+        entityTypeFQN: new FullyQualifiedName(entityType.get('type')),
         entityTypeId: entityType.get('id'),
-        propertyTypeId: removedAbstractTypeId
       });
     }
   }
 
-  handleOnPropertyTypeReorder = (oldIndex :number, newIndex :number) => {
-
-    const { actions, entityType } = this.props;
-
-    if (AuthUtils.isAuthenticated() && AuthUtils.isAdmin()) {
-
-      const propertyTypeIds :List<string> = entityType.get('properties');
-      const idToMove :string = propertyTypeIds.get(oldIndex);
-      const reorderedPropertyTypeIds :List<string> = propertyTypeIds.delete(oldIndex).insert(newIndex, idToMove);
-
-      actions.reorderEntityTypePropertyTypes({
-        entityTypeId: entityType.get('id'),
-        propertyTypeIds: reorderedPropertyTypeIds.toJS()
-      });
-    }
-  }
+  // TODO: uncomment when re-enabling this feature
+  // handleOnPropertyTypeReorder = (oldIndex :number, newIndex :number) => {
+  //
+  //   const { actions, entityType } = this.props;
+  //
+  //   if (AuthUtils.isAuthenticated() && AuthUtils.isAdmin()) {
+  //
+  //     const propertyTypeIds :List<string> = entityType.get('properties');
+  //     const idToMove :string = propertyTypeIds.get(oldIndex);
+  //     const reorderedPropertyTypeIds :List<string> = propertyTypeIds.delete(oldIndex).insert(newIndex, idToMove);
+  //
+  //     actions.reorderEntityTypePropertyTypes({
+  //       entityTypeId: entityType.get('id'),
+  //       propertyTypeIds: reorderedPropertyTypeIds.toJS()
+  //     });
+  //   }
+  // }
 
   handleOnClickDelete = () => {
 
     const { actions, entityType } = this.props;
 
     if (AuthUtils.isAuthenticated() && AuthUtils.isAdmin()) {
-      actions.deleteEntityType(entityType.get('id'));
+      const entityTypeId :?UUID = entityType.get('id');
+      const entityTypeFQN :FQN = new FullyQualifiedName(entityType.get('type'));
+      actions.localDeleteEntityType({ entityTypeFQN, entityTypeId });
     }
   }
 
@@ -126,8 +151,9 @@ class EntityTypeDetailsContainer extends React.Component<Props> {
             highlightOnHover
             maxHeight={500}
             onAbstractTypeRemove={this.handleOnPropertyTypeRemove}
-            onReorder={this.handleOnPropertyTypeReorder}
-            orderable
+            // TODO: uncomment when re-enabling this feature
+            // onReorder={this.handleOnPropertyTypeReorder}
+            // orderable
             showRemoveColumn
             workingAbstractTypeType={AbstractTypes.PropertyType} />
       );
@@ -176,7 +202,7 @@ class EntityTypeDetailsContainer extends React.Component<Props> {
 
   render() {
 
-    const { entityType, propertyTypes, propertyTypesById } = this.props;
+    const { entityType, propertyTypes, propertyTypesIndexMap } = this.props;
 
     if (!entityType || entityType.isEmpty()) {
       return null;
@@ -184,12 +210,12 @@ class EntityTypeDetailsContainer extends React.Component<Props> {
 
     const baseType :string = entityType.get('baseType', '');
 
-    const keyPropertyTypeIds :OrderedSet<string> = entityType.get('key').toOrderedSet();
-    const propertyTypeIds :OrderedSet<string> = entityType.get('properties').toOrderedSet();
+    const keyPropertyTypeIds :OrderedSet<string> = entityType.get('key', List()).toOrderedSet();
+    const propertyTypeIds :OrderedSet<string> = entityType.get('properties', List()).toOrderedSet();
 
     const keyPropertyTypes :List<Map<*, *>> = keyPropertyTypeIds
       .map((propertyTypeId :string) => {
-        const index :number = propertyTypesById.get(propertyTypeId, -1);
+        const index :number = propertyTypesIndexMap.get(propertyTypeId, -1);
         if (index === -1) {
           return Map();
         }
@@ -199,7 +225,7 @@ class EntityTypeDetailsContainer extends React.Component<Props> {
 
     const thePropertyTypes :List<Map<*, *>> = propertyTypeIds
       .map((propertyTypeId :string) => {
-        const index :number = propertyTypesById.get(propertyTypeId, -1);
+        const index :number = propertyTypesIndexMap.get(propertyTypeId, -1);
         if (index === -1) {
           return Map();
         }
@@ -284,26 +310,17 @@ class EntityTypeDetailsContainer extends React.Component<Props> {
   }
 }
 
-function mapStateToProps(state :Map<*, *>) :Object {
+const mapStateToProps = (state :Map<*, *>) :Object => ({
+  propertyTypes: state.getIn(['edm', 'propertyTypes', 'propertyTypes'], List()),
+  propertyTypesIndexMap: state.getIn(['edm', 'propertyTypes', 'propertyTypesIndexMap'], Map()),
+});
 
-  return {
-    propertyTypes: state.getIn(['edm', 'propertyTypes', 'propertyTypes'], List()),
-    propertyTypesById: state.getIn(['edm', 'propertyTypes', 'propertyTypesById'], Map())
-  };
-}
-
-function mapDispatchToProps(dispatch :Function) :Object {
-
-  const actions = {
-    addPropertyTypeToEntityType,
-    deleteEntityType,
-    removePropertyTypeFromEntityType,
-    reorderEntityTypePropertyTypes
-  };
-
-  return {
-    actions: bindActionCreators(actions, dispatch)
-  };
-}
+const mapDispatchToProps = (dispatch :Function) :Object => ({
+  actions: bindActionCreators({
+    localAddPropertyTypeToEntityType: EntityTypesActions.localAddPropertyTypeToEntityType,
+    localDeleteEntityType: EntityTypesActions.localDeleteEntityType,
+    localRemovePropertyTypeFromEntityType: EntityTypesActions.localRemovePropertyTypeFromEntityType,
+  }, dispatch)
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(EntityTypeDetailsContainer);
