@@ -6,210 +6,267 @@
 
 import { List, Map, fromJS } from 'immutable';
 import { Models, Types } from 'lattice';
-import { EntityDataModelApiActionFactory } from 'lattice-sagas';
+import { EntityDataModelApiActions } from 'lattice-sagas';
+import type { FQN, SchemaObject } from 'lattice';
 
-const { FullyQualifiedName, Schema, SchemaBuilder } = Models;
+import Logger from '../../../utils/Logger';
+import {
+  LOCAL_CREATE_SCHEMA,
+  LOCAL_UPDATE_SCHEMA,
+  localCreateSchema,
+  localUpdateSchema,
+} from './SchemasActions';
+import type { IndexMap } from '../Types';
+
+const LOG :Logger = new Logger('SchemasReducer');
+
+const {
+  getEntityDataModel,
+} = EntityDataModelApiActions;
+
+const {
+  FullyQualifiedName,
+  Schema,
+  SchemaBuilder,
+} = Models;
+
 const { ActionTypes } = Types;
-const { createSchema, getAllSchemas, updateSchema } = EntityDataModelApiActionFactory;
 
 const INITIAL_STATE :Map<*, *> = fromJS({
-  actions: {
-    createSchema: Map(),
-    updateSchema: Map()
-  },
-  isCreatingNewSchema: false,
-  isFetchingAllSchemas: false,
-  newlyCreatedSchemaFqn: '',
+  [LOCAL_CREATE_SCHEMA]: { error: false },
+  [LOCAL_UPDATE_SCHEMA]: { error: false },
+  newlyCreatedSchemaFQN: undefined,
   schemas: List(),
-  schemasByFqn: Map()
+  schemasIndexMap: Map(),
 });
 
-export default function propertyTypesReducer(state :Map<*, *> = INITIAL_STATE, action :Object) :Map<*, *> {
+export default function schemasReducer(state :Map<*, *> = INITIAL_STATE, action :Object) :Map<*, *> {
 
   switch (action.type) {
 
-    case createSchema.case(action.type): {
-      return createSchema.reducer(state, action, {
-        REQUEST: () => {
-          // TODO: not ideal. perhaps there's a better way to get access to the trigger action value
-          const seqAction :SequenceAction = (action :any);
-          return state
-            .set('isCreatingNewSchema', true)
-            .set('newlyCreatedSchemaFqn', '')
-            .setIn(['actions', 'createSchema', seqAction.id], fromJS(seqAction));
-        },
+    case getEntityDataModel.case(action.type): {
+      return getEntityDataModel.reducer(state, action, {
         SUCCESS: () => {
 
-          const seqAction :SequenceAction = (action :any);
-          const storedSeqAction :Map<*, *> = state.getIn(['actions', 'createSchema', seqAction.id], Map());
-
-          if (storedSeqAction.isEmpty()) {
+          const seqAction :SequenceAction = action;
+          const responseSchemas :SchemaObject[] = seqAction.value.schemas;
+          if (!responseSchemas || responseSchemas.length === 0) {
+            LOG.error('getEntityDataModel() - Schemas missing', responseSchemas);
             return state;
           }
 
-          const tempSchema :Schema = storedSeqAction.get('value');
-          const newSchemaFqn :string = FullyQualifiedName.toString(tempSchema.fqn);
-          const newSchema :Schema = (new SchemaBuilder())
-            .setFullyQualifiedName(tempSchema.fqn)
-            .setEntityTypes(tempSchema.entityTypes)
-            .setPropertyTypes(tempSchema.propertyTypes)
-            .build();
+          const schemas :List<Map<*, *>> = List().asMutable();
+          const schemasIndexMap :IndexMap = Map().asMutable();
 
-          const iSchema :Map<*, *> = newSchema.asImmutable();
-          const current :List<Map<*, *>> = state.get('schemas', List());
-          const updated :List<Map<*, *>> = current.push(iSchema);
-
-          const currentByFqn :Map<string, number> = state.get('schemasByFqn', Map());
-          const updatedByFqn :Map<string, number> = currentByFqn.set(newSchemaFqn, updated.size - 1);
+          responseSchemas.forEach((s :SchemaObject, index :number) => {
+            try {
+              const schemaFQN :FQN = new FullyQualifiedName(s.fqn);
+              const schema = new SchemaBuilder()
+                .setFullyQualifiedName(schemaFQN)
+                .setEntityTypes(s.entityTypes)
+                .setPropertyTypes(s.propertyTypes)
+                .build();
+              schemas.push(schema.toImmutable());
+              schemasIndexMap.set(schemaFQN, index);
+            }
+            catch (e) {
+              LOG.error('getEntityDataModel()', s);
+              LOG.error('getEntityDataModel()', e);
+            }
+          });
 
           return state
-            .set('newlyCreatedSchemaFqn', newSchemaFqn)
-            .set('schemas', updated)
-            .set('schemasByFqn', updatedByFqn);
-        },
-        FAILURE: () => {
-          // TODO: need to properly handle the failure case
-          return state;
-        },
-        FINALLY: () => {
-          const seqAction :SequenceAction = (action :any);
-          return state
-            .set('isCreatingNewSchema', false)
-            .set('newlyCreatedSchemaFqn', '')
-            .deleteIn(['actions', 'createSchema', seqAction.id]);
-        }
-      });
-    }
-
-    case getAllSchemas.case(action.type): {
-      return getAllSchemas.reducer(state, action, {
-        REQUEST: () => {
-          return state.set('isFetchingAllSchemas', true);
-        },
-        SUCCESS: () => {
-          const seqAction :SequenceAction = (action :any);
-          const schemas :List<Map<*, *>> = fromJS(seqAction.value);
-          const schemasByFqn :Map<string, number> = Map()
-            .withMutations((byIdMap :Map<string, number>) => {
-              schemas.forEach((schema :Map<*, *>, schemaIndex :number) => {
-                byIdMap.set(FullyQualifiedName.toString(schema.get('fqn', Map())), schemaIndex);
-              });
-            });
-          return state
-            .set('schemas', schemas)
-            .set('schemasByFqn', schemasByFqn);
+            .set('schemas', schemas.asImmutable())
+            .set('schemasIndexMap', schemasIndexMap.asImmutable());
         },
         FAILURE: () => {
           return state
             .set('schemas', List())
-            .set('schemasByFqn', Map());
+            .set('schemasIndexMap', Map());
         },
-        FINALLY: () => {
-          return state.set('isFetchingAllSchemas', false);
-        }
       });
     }
 
-    case updateSchema.case(action.type): {
-      return updateSchema.reducer(state, action, {
+    case localCreateSchema.case(action.type): {
+      return localCreateSchema.reducer(state, action, {
         REQUEST: () => {
-          const seqAction :SequenceAction = (action :any);
-          return state.setIn(['actions', 'updateSchema', seqAction.id], fromJS(seqAction));
+          const seqAction :SequenceAction = action;
+          return state
+            .set('newlyCreatedSchemaFQN', undefined)
+            .setIn([LOCAL_CREATE_SCHEMA, seqAction.id], seqAction);
         },
         SUCCESS: () => {
 
-          const seqAction :SequenceAction = (action :any);
-          const storedSeqAction :Map<*, *> = state.getIn(['actions', 'updateSchema', seqAction.id], Map());
-          if (storedSeqAction.isEmpty()) {
-            return state;
+          const seqAction :SequenceAction = action;
+          const storedSeqAction :SequenceAction = state.getIn([LOCAL_CREATE_SCHEMA, seqAction.id]);
+
+          if (storedSeqAction) {
+
+            const storedSchema :Schema = storedSeqAction.value;
+            const schemaFQN :FQN = new FullyQualifiedName(storedSchema.fqn);
+            const newSchema :Schema = new SchemaBuilder()
+              .setFullyQualifiedName(schemaFQN)
+              .setEntityTypes(storedSchema.entityTypes)
+              .setPropertyTypes(storedSchema.propertyTypes)
+              .build();
+
+            const updatedSchemas :List<Map<*, *>> = state
+              .get('schemas')
+              .push(newSchema.toImmutable());
+
+            const newSchemaIndex :number = updatedSchemas.count() - 1;
+            const updatedSchemasIndexMap :IndexMap = state
+              .get('schemasIndexMap')
+              .set(schemaFQN, newSchemaIndex);
+
+            return state
+              .set('newlyCreatedSchemaFQN', schemaFQN)
+              .set('schemas', updatedSchemas)
+              .set('schemasIndexMap', updatedSchemasIndexMap);
           }
 
-          const targetFqn :string = FullyQualifiedName.toString(storedSeqAction.getIn(['value', 'schemaFqn']));
-          const targetIndex :number = state.getIn(['schemasByFqn', targetFqn], -1);
-          if (targetIndex === -1) {
-            return state;
-          }
-
-          const currentSchema :Map<*, *> = state.getIn(['schemas', targetIndex], Map());
-          const currentEntityTypes :List<Map<*, *>> = currentSchema.get('entityTypes', List());
-          const currentPropertyTypes :List<Map<*, *>> = currentSchema.get('propertyTypes', List());
-
-          let workingEntityTypes :List<Map<*, *>> = currentEntityTypes;
-          let workingPropertyTypes :List<Map<*, *>> = currentPropertyTypes;
-
-          const actionEntityTypes :Map<*, *> = storedSeqAction.getIn(['value', 'entityTypes'], Map());
-          const actionEntityTypeIds :List<string> = storedSeqAction.getIn(['value', 'entityTypeIds'], List());
-          const actionPropertyTypes :Map<*, *> = storedSeqAction.getIn(['value', 'propertyTypes'], Map());
-          const actionPropertyTypeIds :List<string> = storedSeqAction.getIn(['value', 'propertyTypeIds'], List());
-
-          if (storedSeqAction.getIn(['value', 'action']) === ActionTypes.ADD) {
-
-            // EntityTypes
-            if (!actionEntityTypeIds.isEmpty() && !actionEntityTypes.isEmpty()) {
-              actionEntityTypes.forEach((actionEntityType :Map<*, *>) => {
-                const matchingIndex :number = workingEntityTypes.findIndex(
-                  (workingEntityType :Map<*, *>) => workingEntityType.get('id') === actionEntityType.get('id')
-                );
-                if (matchingIndex === -1) {
-                  workingEntityTypes = workingEntityTypes.push(actionEntityType);
-                }
-              });
-            }
-
-            // PropertyTypes
-            if (!actionPropertyTypeIds.isEmpty() && !actionPropertyTypes.isEmpty()) {
-              actionPropertyTypes.forEach((actionPropertyType :Map<*, *>) => {
-                const matchingIndex :number = workingPropertyTypes.findIndex(
-                  (workingPropertyType :Map<*, *>) => workingPropertyType.get('id') === actionPropertyType.get('id')
-                );
-                if (matchingIndex === -1) {
-                  workingPropertyTypes = workingPropertyTypes.push(actionPropertyType);
-                }
-              });
-            }
-          }
-          else if (storedSeqAction.getIn(['value', 'action']) === ActionTypes.REMOVE) {
-
-            // EntityTypes
-            if (!actionEntityTypeIds.isEmpty() && !workingEntityTypes.isEmpty()) {
-              actionEntityTypeIds.forEach((entityTypeId :string) => {
-                const targetEntityTypeIndex :number = workingEntityTypes
-                  .findIndex((entityType :Map<*, *>) => entityType.get('id') === entityTypeId);
-                if (targetEntityTypeIndex !== -1) {
-                  workingEntityTypes = workingEntityTypes.delete(targetEntityTypeIndex);
-                }
-              });
-            }
-
-            // PropertyTypes
-            if (!actionPropertyTypeIds.isEmpty() && !workingPropertyTypes.isEmpty()) {
-              actionPropertyTypeIds.forEach((propertyTypeId :string) => {
-                const targetPropertyTypeIndex :number = workingPropertyTypes
-                  .findIndex((propertyType :Map<*, *>) => propertyType.get('id') === propertyTypeId);
-                if (targetPropertyTypeIndex !== -1) {
-                  workingPropertyTypes = workingPropertyTypes.delete(targetPropertyTypeIndex);
-                }
-              });
-            }
-          }
-          else {
-            return state;
-          }
-
-          const updatedSchema :Map<*, *> = currentSchema
-            .set('entityTypes', workingEntityTypes)
-            .set('propertyTypes', workingPropertyTypes);
-
-          return state.setIn(['schemas', targetIndex], updatedSchema);
+          return state;
         },
         FAILURE: () => {
-          // TODO: need to properly handle the failure case
+          const seqAction :SequenceAction = action;
+          const storedSeqAction :SequenceAction = state.getIn([LOCAL_CREATE_SCHEMA, seqAction.id]);
+          if (storedSeqAction) {
+            // TODO: we need a better pattern for setting and handling errors
+            return state
+              .set('newlyCreatedSchemaFQN', undefined)
+              .setIn([LOCAL_CREATE_SCHEMA, 'error'], true);
+          }
           return state;
         },
         FINALLY: () => {
-          const seqAction :SequenceAction = (action :any);
-          return state.deleteIn(['actions', 'updateSchema', seqAction.id]);
-        }
+          const seqAction :SequenceAction = action;
+          return state.deleteIn([LOCAL_CREATE_SCHEMA, seqAction.id]);
+        },
+      });
+    }
+
+    case localUpdateSchema.case(action.type): {
+      return localUpdateSchema.reducer(state, action, {
+        REQUEST: () => {
+          const seqAction :SequenceAction = action;
+          return state.setIn([LOCAL_UPDATE_SCHEMA, seqAction.id], seqAction);
+        },
+        SUCCESS: () => {
+
+          const seqAction :SequenceAction = action;
+          const storedSeqAction :SequenceAction = state.getIn([LOCAL_UPDATE_SCHEMA, seqAction.id]);
+          if (storedSeqAction) {
+
+            const {
+              actionType,
+              entityTypeIds,
+              entityTypes,
+              propertyTypeIds,
+              propertyTypes,
+              schemaFQN,
+            } = storedSeqAction.value;
+
+            const targetIndex :number = state.getIn(['schemasIndexMap', schemaFQN], -1);
+            if (targetIndex === -1) {
+              LOG.error('Schema does not exist in store', schemaFQN);
+              return state;
+            }
+            const target :Map<*, *> = state.getIn(['schemas', targetIndex], Map());
+            const targetFQN :FQN = new FullyQualifiedName(target.get('fqn', Map()));
+            if (targetFQN.toString() !== schemaFQN.toString()) {
+              LOG.error('EntityType does not match fqn', schemaFQN);
+              return state;
+            }
+
+            let newState :Map<*, *> = state;
+            if (actionType === ActionTypes.ADD) {
+
+              // EntityTypes
+              if (entityTypes && entityTypes.length > 0) {
+                entityTypes.forEach((entityTypeToAdd :Map<*, *>) => {
+                  const index :number = target.get('entityTypes').findIndex(
+                    (entityType :Map<*, *>) => entityType.get('id') === entityTypeToAdd.get('id')
+                  );
+                  // only add if not in the list
+                  if (index === -1) {
+                    newState = newState.setIn(
+                      ['schemas', targetIndex, 'entityTypes'],
+                      newState.getIn(['schemas', targetIndex, 'entityTypes']).push(entityTypeToAdd)
+                    );
+                  }
+                });
+              }
+
+              // PropertyTypes
+              if (propertyTypes && propertyTypes.length > 0) {
+                propertyTypes.forEach((propertyTypeToAdd :Map<*, *>) => {
+                  const index :number = target.get('propertyTypes').findIndex(
+                    (propertyType :Map<*, *>) => propertyType.get('id') === propertyTypeToAdd.get('id')
+                  );
+                  // only add if not in the list
+                  if (index === -1) {
+                    newState = newState.setIn(
+                      ['schemas', targetIndex, 'propertyTypes'],
+                      newState.getIn(['schemas', targetIndex, 'propertyTypes']).push(propertyTypeToAdd)
+                    );
+                  }
+                });
+              }
+            }
+            else if (actionType === ActionTypes.REMOVE) {
+
+              // EntityTypes
+              if (entityTypeIds && entityTypeIds.length > 0) {
+                entityTypeIds.forEach((entityTypeId :UUID) => {
+                  const targetEntityTypeIndex :number = target.get('entityTypes').findIndex(
+                    (entityType :Map<*, *>) => entityType.get('id') === entityTypeId
+                  );
+                  if (targetEntityTypeIndex !== -1) {
+                    newState = newState.deleteIn(['schemas', targetIndex, 'entityTypes', targetEntityTypeIndex]);
+                  }
+                  else {
+                    LOG.error('EntityType not found in Schema', entityTypeId);
+                  }
+                });
+              }
+
+              // PropertyTypes
+              if (propertyTypeIds && propertyTypeIds.length > 0) {
+                propertyTypeIds.forEach((propertyTypeId :UUID) => {
+                  const targetPropertyTypeIndex :number = target.get('propertyTypes').findIndex(
+                    (propertyType :Map<*, *>) => propertyType.get('id') === propertyTypeId
+                  );
+                  if (targetPropertyTypeIndex !== -1) {
+                    newState = newState.deleteIn(['schemas', targetIndex, 'propertyTypes', targetPropertyTypeIndex]);
+                  }
+                  else {
+                    LOG.error('PropertyType not found in Schema', propertyTypeId);
+                  }
+                });
+              }
+            }
+            else {
+              LOG.error('invalid ActionType', actionType);
+            }
+
+            return newState;
+          }
+
+          return state;
+        },
+        FAILURE: () => {
+          const seqAction :SequenceAction = action;
+          const storedSeqAction :SequenceAction = state.getIn([LOCAL_UPDATE_SCHEMA, seqAction.id]);
+          if (storedSeqAction) {
+            // TODO: we need a better pattern for setting and handling errors
+            return state.setIn([LOCAL_UPDATE_SCHEMA, 'error'], true);
+          }
+          return state;
+        },
+        FINALLY: () => {
+          const seqAction :SequenceAction = action;
+          return state.deleteIn([LOCAL_UPDATE_SCHEMA, seqAction.id]);
+        },
       });
     }
 
