@@ -2,7 +2,7 @@
  * @flow
  */
 
-import createHistory from 'history/createHashHistory';
+import { createHashHistory as createHistory } from 'history';
 
 export default function initializeRouterHistory() :any {
 
