@@ -6,11 +6,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import LatticeAuth from 'lattice-auth';
+import { ConnectedRouter } from 'connected-react-router/immutable';
 import { normalize } from 'polished';
 import { Provider } from 'react-redux';
 import { Route, Switch } from 'react-router-dom';
-import { ConnectedRouter } from 'react-router-redux';
-import { injectGlobal } from 'styled-components';
+import { createGlobalStyle } from 'styled-components';
 
 import AppContainer from './containers/app/AppContainer';
 import initializeReduxStore from './core/redux/ReduxStore';
@@ -29,10 +29,12 @@ const {
 } = LatticeAuth;
 
 /* eslint-disable */
-injectGlobal`${normalize()}`;
+const NormalizeCSS = createGlobalStyle`
+  ${normalize()}
+`;
 
 // TODO: define style defaults and themes
-injectGlobal`
+const GlobalStyle = createGlobalStyle`
   html,
   body {
     background-color: #f9fcff;
@@ -43,16 +45,12 @@ injectGlobal`
   }
 
   * {
-    -webkit-box-sizing: border-box;
-       -moz-box-sizing: border-box;
-            box-sizing: border-box;
+    box-sizing: border-box;
   }
 
-  *:before,
-  *:after {
-    -webkit-box-sizing: border-box;
-       -moz-box-sizing: border-box;
-            box-sizing: border-box;
+  *::before,
+  *::after {
+    box-sizing: border-box;
   }
 
   #app {
@@ -85,12 +83,16 @@ const APP_ROOT_NODE = document.getElementById('app');
 if (APP_ROOT_NODE) {
   ReactDOM.render(
     <Provider store={reduxStore}>
-      <ConnectedRouter history={routerHistory}>
-        <Switch>
-          <AuthRoute exact strict path={Routes.LOGIN} />
-          <Route path={Routes.ROOT} component={AppContainer} />
-        </Switch>
-      </ConnectedRouter>
+      <>
+        <ConnectedRouter history={routerHistory}>
+          <Switch>
+            <AuthRoute exact strict path={Routes.LOGIN} />
+            <Route path={Routes.ROOT} component={AppContainer} />
+          </Switch>
+        </ConnectedRouter>
+        <NormalizeCSS />
+        <GlobalStyle />
+      </>
     </Provider>,
     APP_ROOT_NODE
   );
